+++ vmware_horizon.py
@@ -1656,6 +1656,87 @@ class Inventory:
             else:
                 return response.json()
 
+    def check_application_name_availability(self,application_name: str)-> dict:
+        """Checks if the given name is available for application pool creation.
+
+        Requires the name of the application to test as string
+        Available for Horizon 8 2103 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        data["name"] = application_name
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/application-pools/action/check-name-availability', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def check_desktop_pool_name_availability(self,desktop_pool_name: str)-> dict:
+        """Checks if the given name is available for desktop pool creation.
+
+        Requires the name of the desktop pool to test as string
+        Available for Horizon 8 2103 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        data["name"] = desktop_pool_name
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/desktop-pools/action/check-name-availability', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def check_farm_name_availability(self,farm_name: str)-> dict:
+        """Checks if the given name is available for farm creation.
+
+        Requires the name of the farm to test as string
+        Available for Horizon 8 2103 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        data["name"] = farm_name
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/inventory/v1/farms/action/check-name-availability', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
 class Monitor:
     def __init__(self, url: str, access_token: dict):
         """Default object for the monitor class used for the monitoring of the various VMware Horiozn services."""
@@ -1958,11 +2039,8 @@ class Monitor:
         """Lists metrics of desktop pools (except RDS desktop pools).
 
         Available for Horizon 8 2012 and later."""
-        headers = self.access_token
-        headers["Content-Type"] = 'application/json'
-        data = desktop_pool_ids
-        json_data = json.dumps(data)
-        response = requests.post(f'{self.url}/rest/inventory/v1/rds-servers/action/recover', verify=False,  headers=headers, data = json_data)
+        urlstring='&ids='.join(desktop_pool_ids)
+        response = requests.get(f'{self.url}/rest/monitor/v1/desktop-pools/metrics?ids={urlstring}', verify=False,  headers=self.access_token)
         if response.status_code == 400:
             error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
@@ -2304,6 +2382,37 @@ class Config:
             else:
                 return response.status
 
+    def get_local_access_groups(self) -> list:
+        """Lists all local access groups.
+
+        Available for Horizon 8 2103 and later."""
+        response = requests.get(f'{self.url}/rest/config/v1/local-access-groups', verify=False,  headers=self.access_token)
+        if response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def get_local_access_group(self,local_access_group_id: str) -> dict:
+        """Retrieves a local access group.
+
+        Requires the id of an local access group as string
+        Available for Horizon 8 2103 and later."""
+        response = requests.get(f'{self.url}/rest/config/v1/local-access-groups/{local_access_group_id}', verify=False,  headers=self.access_token)
+        if response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
 class External:
     def __init__(self, url: str, access_token: dict):
         """Default object for the External class for resources that are external to Horizon environment."""
@@ -2550,6 +2659,27 @@ class External:
             else:
                 return response.json()
 
+    def get_datastore_clusters(self, vcenter_id : str, host_or_cluster_id:str ) -> list:
+        """Lists all the datastore clusters from the vCenter for the given host or cluster.
+
+        Requires host_or_cluster_id and vcenter_id
+        Available for Horizon 8 2103 and later."""
+        response = requests.get(f'{self.url}/rest/external/v1/datastore-clusters?host_or_cluster_id={host_or_cluster_id}&vcenter_id={vcenter_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 404:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
     def get_datastore_paths(self, vcenter_id : str, datastore_id:str ) -> list:
         """Lists all the folder paths within a Datastore from vCenter.
 
@@ -2691,6 +2821,69 @@ class External:
                 raise "Error: " + str(e)
             else:
                 return response.json()
+
+    def compute_datastore_requirement(self,desktop_or_farm_id: str, user_assignment : str, vcenter_id : str, pool_size : int, source : str, type : str, base_snapshot_id: str="",base_vm_id: str="", use_separate_replica_and_os_disk : bool=False, use_vsan : bool = False, vm_template_id : str = "") -> list:
+        """Creates Instant Clone Domain Account
+
+        Requirements:
+        base_snapshot_id : str - Required when source is INSTANT_CLONE
+        base_vm_id : string - Required when source is INSTANT_CLONE
+        desktop_or_farm_id: string
+        pool_size : int32
+        source : string - Required to be either FULL_CLONE or INSTANT_CLONE
+        type : string - Required to be DESKTOP_POOL or FARM
+        use_separate_replica_and_os_disk : boolean - Ignored for FULL_CLONE or when vSAN is used, defaults to False
+        use_vsan : boolean - defaults to False
+        user_assignment : string - Required to be DEDICATED or FLOATING
+        vcenter_id : string
+        vm_template_id : string - Required when source is FULL_CLONE
+
+        Available for Horizon 8 2103 and later."""
+        if source !="FULL_CLONE" and source !="INSTANT_CLONE":
+            raise Exception(f"Error: source has to be either FULL_CLONE or INSTANT_CLONE")
+        if type !="DESKTOP_POOL" and type !="FARM":
+            raise Exception(f"Error: type has to be either DESKTOP_POOL or FARM")
+        if user_assignment !="DEDICATED" and user_assignment !="FLOATING":
+            raise Exception(f"Error: user_assignment has to be either DEDICATED or FLOATING")
+        if source == "FULL_CLONE" and vm_template_id == "":
+            raise Exception(f"Error: When type is FULL_CLONE vm_template_id is required ")
+        if source == "INSTANT_CLONE" and (base_vm_id == "" or base_snapshot_id == ""):
+            raise Exception(f"Error: When type is INSTANT_CLONE both base_vm_id and base_snapshot_id are required ")
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        if source == "INSTANT_CLONE":
+            data["base_snapshot_id"] = base_snapshot_id
+            data["base_vm_id"] = base_vm_id
+        data["id"] = desktop_or_farm_id
+        data["pool_size"] = pool_size
+        data["source"] = source
+        data["type"] = type
+        data["use_separate_replica_and_os_disk"] = use_separate_replica_and_os_disk
+        data["use_vsan"] = use_vsan
+        data["user_assignment"] = user_assignment
+        data["vcenter_id"] = vcenter_id
+        if type == "FULL_CLONE":
+            data["vm_template_id"] = vm_template_id
+        json_data = json.dumps(data)
+        response = requests.post(f'{self.url}/rest/external/v1/datastores/action/compute-requirements', verify=False,  headers=headers, data=json_data)
+        if response.status_code == 400:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 409:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+
 
 class Entitlements:
     def __init__(self, url: str, access_token: dict):